Adds handling of intrinsic for array handling Adds handling of arbitrary type for array load stores Un XFAILs array and mapping testcases

+++ tests/e2e/feature/array.sol
@@ -5,16 +5,26 @@ This testcase targets storage vaiable in same slot
  */
 
 // RUN: pyul %s -o %t --project-dir %S | FileCheck %s
-// XFAIL: *
 
 pragma solidity ^0.8.10;
 
 contract ArrayTest {
-    function readArray(uint256[10] memory array, uint256 index) external view returns (uint256){
+    function readArray(uint32[10] memory array, uint256 index) external view returns (uint32){
         return array[index];
     }
 
-    function writeArray(uint256[10] memory array, uint256 index, uint256 value) external {
+    function writeArray(uint32[10] memory array, uint256 index, uint32 value) external {
         array[index] = value;
     }
 }
+
+//CHECK: define i256 @fun_readArray_{{[0-9]+}}(i256* %{{.+, .+, .+}})
+//CHECK: getelementptr [0 x {{.*}}], [0 x {{.*}}]*
+//CHECK: arr_load{{.*}} = load i256, i256*
+//CHECK: cleaned_up{{.*}} = and i256
+
+
+//CHECK: define void @fun_writeArray_{{[0-9]+}}(i256* %{{.*, .+, .+, .+}})
+//CHECK: getelementptr [0 x {{.*}}], [0 x {{.*}}]*
+//CHECK: cleaned_up{{.*}} = and i256
+//CHECK: store i256 {{.*}}, i256*
